closes #859 get all path doc strings tests in same place
