+++ src/a18_etl_toolbox/test/_util/test_a18_path.py
@@ -1,4 +1,7 @@
+from inspect import getdoc as inspect_getdoc
+from platform import system as platform_system
 from src.a00_data_toolbox.file_toolbox import create_path
+from src.a09_pack_logic.test._util.a09_str import believer_name_str
 from src.a18_etl_toolbox.a18_path import (
     STANCE0001_FILENAME,
     create_stance0001_path,
@@ -52,3 +55,32 @@ def test_create_stance0001_path_ReturnsObj():
     # THEN
     expected_stance000001_path = create_path(output_dir, STANCE0001_FILENAME)
     assert gen_stance0001_xlsx_path == expected_stance000001_path
+
+
+LINUX_OS = platform_system() == "Linux"
+
+
+def test_create_stances_dir_path_HasDocString():
+    # ESTABLISH
+    doc_str = create_stances_dir_path(belief_mstr_dir="belief_mstr_dir")
+    doc_str = f"Returns path: {doc_str}"
+    # WHEN / THEN
+    assert LINUX_OS or inspect_getdoc(create_stances_dir_path) == doc_str
+
+
+def test_create_stances_believer_dir_path_HasDocString():
+    # ESTABLISH
+    doc_str = create_stances_believer_dir_path(
+        belief_mstr_dir="belief_mstr_dir", believer_name=believer_name_str()
+    )
+    doc_str = f"Returns path: {doc_str}"
+    # WHEN / THEN
+    assert LINUX_OS or inspect_getdoc(create_stances_believer_dir_path) == doc_str
+
+
+def test_create_stance0001_path_HasDocString():
+    # ESTABLISH
+    doc_str = create_stance0001_path(output_dir="output_dir")
+    doc_str = f"Returns path: {doc_str}"
+    # WHEN / THEN
+    assert LINUX_OS or inspect_getdoc(create_stance0001_path) == doc_str
